Improve usability for Command Line tool. Change ulong datatypes to int. Write some general troubleshooting documents. Fix small build issue.

+++ LoopCheckTool.Console/LibraryException.cs
@@ -8,21 +8,21 @@ namespace LoopCheckTool.Console
 {
     public class LibraryException : Exception
     {
-        public ulong AffectedRow { get; }
+        public int AffectedRow { get; }
         public IDictionary<string, string> RowData { get; }
-        public LibraryException(ulong affectedRow, IDictionary<string, string> rowData) : base()
+        public LibraryException(int affectedRow, IDictionary<string, string> rowData) : base()
         {
             AffectedRow = affectedRow;
             RowData = rowData;
         }
 
-        public LibraryException(string message, ulong affectedRow, IDictionary<string, string> rowData) : base(message)
+        public LibraryException(string message, int affectedRow, IDictionary<string, string> rowData) : base(message)
         {
             AffectedRow = affectedRow;
             RowData = rowData;
         }
 
-        public LibraryException(string message, Exception innerException, ulong affectedRow, IDictionary<string, string> rowData) : base(message, innerException)
+        public LibraryException(string message, Exception innerException, int affectedRow, IDictionary<string, string> rowData) : base(message, innerException)
         {
             AffectedRow = affectedRow;
             RowData = rowData;
